% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package10883408-package.R
\docType{package}
\name{package10883408-package}
\alias{package10883408}
\alias{package10883408-package}
\title{package10883408: Article Data Analysis and Visualization with Package}
\description{
The package_10883408 made for text analysis. It is perfect for researchers because it provides essential features for analysing word frequencies in datasets. It provides below features.
}
\keyword{internal}
