% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zipfs_law.R
\name{plot_zipfs_law}
\alias{plot_zipfs_law}
\title{Plot Zipf's Law for Article Data}
\usage{
plot_zipfs_law(article_data, articles, journal)
}
\arguments{
\item{article_data}{A dataframe containing the articles and their metadata.}

\item{articles}{The column name in `article_data` that contains the article text.}

\item{journal}{The column name in `article_data` that contains the journal names.}
}
\value{
A list containing the processed article data, regression analysis summary, and the ggplot object.
}
\description{
This function applies Zipf's law to a given dataset of articles. It tokenizes the words in the articles,
performs a count and a TF-IDF (Term Frequency-Inverse Document Frequency) calculation, then uses linear
regression to analyze the relationship between the ranks of words and their frequencies. The result is
a log-log plot illustrating Zipf's law, with different colors for each journal.
}
\examples{
# Example usage:
plot_zipfs_law(article_data ,"articles","journal")
}
